+++ include/stdexec/coroutine.hpp
@@ -15,4 +15,4 @@
  */
 #pragma once
 
-#include "__detail/__coroutine.hpp"
+#include "__detail/__awaitable.hpp"
